tests/torch/requirements.txt
yattag>=1.14.0
prettytable>=2.0.0
onnx>=1.8.0
onnxruntime==1.6.0
pytest-mock>=3.3.1
pytest-dependency>=0.5.1
virtualenv
# Ticket 69520
pyparsing<3.0

# Required for search_buidling_blocks tests
transformers==4.23.1

# Required for movement_sparsity tests
datasets==2.6.2
evaluate==0.3.0